requirements.txt
anthropic==0.25.8
gradio==4.29.0
numpy==1.25.2
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.31.0
resemble==1.4.2

